feature(@nestjs/cqrs) Added the subject as the constructor in the default pubsub, allowing it to skip checks at publishing time.
Made the event bus check if the publisher implements an optional "publishAll" method, and calls it if available, or fallback to iterating and calling publish for each event.

The "publish" and "publishAll" methods return the results from the publisher, which is still "any" as before, and shouldn't be relied on with the default, but allows extensions to override them AND rely on the return value if they have to (f.e. to resolve a promise about the event being dispatched).