fix(decorators): check for own metadata in command/query handler
When there is two commands/queries that one inherits from the other and both of them have registered handlers, then metadata for them is defined only for parent and not for child. This causes the same ID to be generated for both commands/queries, sometimes resulting in wrong handler picking up the message.